bash: remove parallelization patch
This patch is problematic in two ways, first with the mis-spelled
nproc (as nprocs), it causes the -j option to not have a value, this
causes make to not limit the number of jobs that can run in parallel.
Secondly, the bash build seems to have a race condition when a large
number of jobs (over 60) are run.

Since we want a more stable build, let's remove this patch.

Change-Id: I5d3e3f496c067fa837090cfec52ddb7aca8873f0
Story: 2003358
Task: 24408
Signed-off-by: Saul Wold <sgw@linux.intel.com>