bash: enable to log the shell command
After merging the upversion commit
 https://review.opendev.org/c/starlingx/compile/+/771784,
the new version add a condition check "syslog_history" variable to
enable/disable the syslog of bash command.
If the syslog_history shopt variable is unset as default, the shell
commands won't be logged.

Now we always enable it, since the commands run by every user in a
login shell need to be logged to /var/log/bash.log. This is very
important as an aid in troubleshooting and debugging issues.

Closes-Bug: #1917864

Change-Id: I4aa2f49a0ea4c54a0e836b8ccb33bcc173653252
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>