tools: Add dwarves package version 1.22
The kernel build, when CONFIG_DEBUG_INFO_BTF=y is set, requires the
pahole helper program, with a version >= 1.16. pahole is shipped in the
dwarves package, but CentOS 7 repositories do not include a sufficiently
recent version of the dwarves package.

Hence, to able to build the kernel with the aforementioned configuration
option, this commit adds the dwarves package, version 1.22, to the
compiler layer. 1.22 is the latest released version as of this writing.

The dwarves package spec file introduced by this patch was acquired and
adapted from the 1.22 release's tar archive, available at:
  https://fedorapeople.org/~acme/dwarves/dwarves-1.22.tar.bz2

Verification: pahole was confirmed to build successfully with this
commit in a monolithic StarlingX build environment. The same build
environment was successfully used to build a kernel with
CONFIG_DEBUG_INFO_BTF=y set.

Story: 2008921
Task: 43662

Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Change-Id: Ied2400c2c0a7ada6f974ac3f4299d1bd22b4f231