Replace golang 1.16.6 with 1.16.12
In order to add kubernetes 1.21.8 we need to add support for
golang 1.16.12.  We're going to be removing kubernetes 1.21.3 so
we no longer need golang 1.16.6.

The ppc64le-vdso-fix.patch patch was merged upstream and so is
no longer applicable.

No direct testing is possible, but it is tested indirectly by
building and testing Kubernetes 1.21.8.

Partial-Bug: 1957994
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: I25d18ee9c7e6a2a2aaa72c896b4687768804fa30